# tests/openapi.json
{
  "components": {
    "schemas": {
...
      "AsyncTranslationRequest": {
        "additionalProperties": false,
        "description": "Async submission accepts inline text or a token-bound text artifact.",
        "properties": {
          "api_base": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Base"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "channel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          },
          "dual_translate": {
            "default": false,
            "title": "Dual Translate",
            "type": "boolean"
          },
          "glossary": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "maxItems": 1000,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Glossary"
          },
          "glossary_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Glossary Text"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "second_api_base": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Second Api Base"
          },
          "second_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Second Api Key"
          },
          "second_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Second Model"
          },
          "sliding_window_words": {
            "default": 80,
            "maximum": 2000.0,
            "minimum": 0.0,
            "title": "Sliding Window Words",
            "type": "integer"
          },
          "target_language": {
            "default": "English",
            "maxLength": 80,
            "minLength": 1,
            "title": "Target Language",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "text_artifact_id": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Artifact Id"
          },
          "text_artifact_token": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Artifact Token"
          }
        },
        "title": "AsyncTranslationRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NllbRequest": {
        "additionalProperties": false,
        "properties": {
          "target_language": {
            "default": "English",
            "title": "Target Language",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          }
        },
        "title": "NllbRequest",
        "type": "object"
      },
...
      "TranscriptionConfigUpdate": {
        "additionalProperties": false,
        "properties": {
          "api_base": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Base"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "engine": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranscriptionEngineType"
              },
              {
                "type": "null"
              }
            ]
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "transcription_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcription Api Key"
          }
        },
        "title": "TranscriptionConfigUpdate",
        "type": "object"
      },
      "TranscriptionEngineType": {
        "enum": [
          "api",
          "whisper_api",
          "local",
          "whisper_local",
          "faster_whisper",
          "faster-whisper",
          "auto"
        ],
        "title": "TranscriptionEngineType",
        "type": "string"
      },
      "TranslationRequest": {
        "additionalProperties": false,
        "properties": {
          "api_base": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Base"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "dual_translate": {
            "default": false,
            "title": "Dual Translate",
            "type": "boolean"
          },
          "glossary": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "maxItems": 1000,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Glossary"
          },
          "glossary_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Glossary Text"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "second_api_base": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Second Api Base"
          },
          "second_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Second Api Key"
          },
          "second_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Second Model"
          },
          "sliding_window_words": {
            "default": 80,
            "maximum": 2000.0,
            "minimum": 0.0,
            "title": "Sliding Window Words",
            "type": "integer"
          },
          "target_language": {
            "default": "Spanish",
            "maxLength": 80,
            "minLength": 1,
            "title": "Target Language",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "text_artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Artifact Id"
          },
          "text_artifact_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Artifact Token"
          }
        },
        "title": "TranslationRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "OmniScribe PDF OCR and document processing API",
    "title": "OmniScribe API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/config/transcription": {
      "get": {
        "description": "Return the current transcription configuration.\n\nThe config controls the active model, language hint, and any\nbackend-specific options (e.g. Whisper ``beam_size``). The\nFlutter client reads this on screen mount so the UI matches\nthe persisted state.",
        "operationId": "get_transcription_config_api_config_transcription_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Transcription Config",
        "tags": [
          "transcribe"
        ]
      },
      "post": {
        "description": "Persist an updated transcription configuration.\n\nValidation failures (``TranscribeError``) propagate so the\nframework exception handler maps them to the standard error\nenvelope; successful updates echo the new config back to the\ncaller.",
        "operationId": "update_transcription_config_api_config_transcription_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscriptionConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Transcription Config",
        "tags": [
          "transcribe"
        ]
      }
    },
...
    "/api/translate": {
      "post": {
        "description": "Synchronously translate inline text or a referenced artifact.\n\nThe route accepts EITHER an inline ``text`` payload OR a\n``text_artifact_id`` + ``text_artifact_token`` pair. Both empty\nreturns the standard 400 envelope so the client sees a uniform\nerror rather than a 500 from a downstream empty-input crash.",
        "operationId": "translate_api_translate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Translate",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/async": {
      "post": {
        "description": "Enqueue an async translation job.\n\nThe artifact pair is optional-with-bounds on the schema, so a\nmissing pair never 422s; the route owns the 400 contract. The\nresponse is the same shape returned by the in-process\nJobQueue submit path (``job_id`` + ``status``).",
        "operationId": "translate_async_api_translate_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AsyncTranslationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Translate Async",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/nllb": {
      "post": {
        "description": "Run a one-shot NLLB-200 translation on the local model.\n\nKept as a separate route from :func:`translate` because the\nNLLB backend has a different prompt surface (no glossary\ninjection, no async queue) and is exercised by clients that\nspecifically want raw NLLB behaviour without tree-aware\npostprocessing.",
        "operationId": "translate_nllb_api_translate_nllb_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NllbRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Translate Nllb",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/result/{job_id}": {
      "get": {
        "description": "Fetch the token-bound translation result for a completed job.\n\nMissing/wrong token, unknown job, or still-running job all map\nto the same 404 (no existence leak \u2014 see C-3/H-3 semantics in\n``docs/SECURITY.md``).",
        "operationId": "translate_result_api_translate_result__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "default": "",
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Artifact-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Artifact-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Translate Result",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/status/{job_id}": {
      "get": {
        "description": "Return the current status of an async translation job.\n\nAn unknown ``job_id`` is mapped to 404 rather than 200-with-null\nso the client can treat it as a terminal failure to display.",
        "operationId": "translation_status_api_translate_status__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Translation Status",
        "tags": [
          "translate"
        ]
      }
    },
...
  }
}